pyproject.toml
[project]
name = "strava-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.46",
    "stravalib>=2.4",
    "streamlit>=1.54.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "folium>=0.20.0",
    "matplotlib>=3.10.8",
    "pytest>=9.0.2",
    "seaborn>=0.13.2",
]
